backend/email_management/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
firebase-admin
python-dotenv
google-generativeai
openai
cryptography
requests
supabase
